Add avatar API for users without avatars. Refactor
